making teaching-materials band button to work, fixing bug
