+++ components/LessonSection/index.js
@@ -70,7 +70,7 @@ const LessonSection = ({ index, section, _sectionDots, ForGrades }) => {
   let compProps = { ...section, _sectionDots };
   const parentId = `${section.SectionTitle}-parent-${index}`;
 
-  if((typeof compProps.SectionTitle === "string") && compProps.SectionTitle.includes("Teaching Materials")){
+  if ((typeof compProps.SectionTitle === "string") && compProps.SectionTitle.includes("Teaching Materials")) {
     compProps = {
       ...compProps,
       ForGrades: ForGrades,
